backend/requirements.txt
fastapi
uvicorn
pymysql
fastapi[standard]
pydantic
passlib[bcrypt]
pyjwt
mysql-connector-python